add multi-cp support to cleanup command
