+++ cmd/cleanup/flag.go
@@ -6,32 +6,41 @@ import (
 )
 
 const (
+	flagConfig     = "config"
 	flagClusterID  = "cluster"
 	flagKubeconfig = "kubeconfig"
-	flagEndpoint   = "endpoint"
-	flagInCluster  = "in-cluster"
-	flagToken      = "token"
+	flagProvider   = "provider"
 )
 
 type flag struct {
 	ClusterID  string
+	Config     string
 	Kubeconfig string
-	Endpoint   string
-	InCluster  bool
-	Token      string
+	Provider   string
 }
 
 func (f *flag) Init(cmd *cobra.Command) {
 	cmd.Flags().StringVarP(&f.ClusterID, flagClusterID, "c", "", `The ID of the cluster to delete.`)
+	cmd.Flags().StringVarP(&f.Config, flagConfig, "g", "", `The path to the file containing API endpoints and tokens for each provider.`)
 	cmd.Flags().StringVarP(&f.Kubeconfig, flagKubeconfig, "k", "", `The path to the kubeconfig for the control plane.`)
-	cmd.Flags().StringVarP(&f.Endpoint, flagEndpoint, "n", "", `The endpoint of the target control plane's API.`)
-	cmd.Flags().BoolVarP(&f.InCluster, flagInCluster, "i", false, `True if this program is running in a Kubernetes cluster and should communicate with the API via the injected service account token.`)
-	cmd.Flags().StringVarP(&f.Token, flagToken, "t", "", `The token used to authenticate with the GS API.`)
+	cmd.Flags().StringVarP(&f.Provider, flagProvider, "k", "", `The path to the kubeconfig for the control plane.`)
 }
 
 func (f *flag) Validate() error {
-	if f.Kubeconfig == "" && !f.InCluster || f.Kubeconfig != "" && f.InCluster {
-		return microerror.Maskf(invalidFlagError, "--%s and --%s are mutually exclusive", flagKubeconfig, flagInCluster)
+	if f.ClusterID == "" {
+		return microerror.Maskf(invalidFlagError, "--%s is required", flagClusterID)
+	}
+
+	if f.Config == "" {
+		return microerror.Maskf(invalidFlagError, "--%s is required", flagConfig)
+	}
+
+	if f.Kubeconfig == "" {
+		return microerror.Maskf(invalidFlagError, "--%s is required", flagKubeconfig)
+	}
+
+	if f.Provider == "" {
+		return microerror.Maskf(invalidFlagError, "--%s is required", flagProvider)
 	}
 
 	return nil
