+++ pkg/key/gsctl.go
@@ -1,0 +1,7 @@
+package key
+
+const (
+	CreationResultError            = "error"
+	CreationResultCreatedWithError = "created-with-errors"
+	DeletionResultScheduled        = "deletion scheduled"
+)
